Implemented changes suggested during code review:
 * Migrated from using float to double;
 * Renamed DtlPointsEstimationDialog to DtlPointsEstimationFragment (+ layout name changed too);
 * Parcelable-related code re-arranged for better readability;
 * 'Points Estimation' error handling passed to super (show snackbar by default);
 * Redundant event for clicking calculation removed - using presenter method instead;
 * Resources for showing input error moved to strings.xml + supplementary code changes.

+++ app/src/main/java/com/worldventures/dreamtrips/core/api/DtlApi.java
@@ -17,5 +17,5 @@ public interface DtlApi {
     DtlPlacesHolder getDtlPlaces(@Path("id") int locationId);
 
     @GET("/api/dtl/places/{id}/points")
-    EstimationPointsHolder getDtlPlacePointsEstimation(@Path("id") int id, @Query("price") float price);
+    EstimationPointsHolder getDtlPlacePointsEstimation(@Path("id") int id, @Query("price") double price);
 }
