+++ app/src/main/java/com/worldventures/dreamtrips/modules/dtl/api/GetDtlPlacePointsEstimationQuery.java
@@ -1,18 +1,18 @@
 package com.worldventures.dreamtrips.modules.dtl.api;
 
-public class GetDtlPlacePointsEstimationQuery extends DtlRequest<Float> {
+public class GetDtlPlacePointsEstimationQuery extends DtlRequest<Double> {
 
     private int id;
-    private float price;
+    private double price;
 
-    public GetDtlPlacePointsEstimationQuery(int id, float price) {
-        super(Float.class);
+    public GetDtlPlacePointsEstimationQuery(int id, double price) {
+        super(Double.class);
         this.id = id;
         this.price = price;
     }
 
     @Override
-    public Float loadDataFromNetwork() {
+    public Double loadDataFromNetwork() {
         return getService().getDtlPlacePointsEstimation(id, price).getPoints();
     }
 }
